Split create-storybook package from toolbox
