+++ code/lib/create-storybook/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@storybook/toolbox",
+  "name": "create-storybook",
   "version": "8.2.0-beta.0",
   "description": "Storybook's Toolbox",
   "homepage": "https://github.com/storybookjs/storybook/tree/next/code/lib/toolbox",
@@ -39,28 +39,16 @@
     "prep": "node --loader ../../../scripts/node_modules/esbuild-register/loader.js -r ../../../scripts/node_modules/esbuild-register/register.js ../../../scripts/prepare/bundle.ts"
   },
   "dependencies": {
-    "@babel/core": "^7.24.4",
-    "@babel/types": "^7.24.0",
     "@ndelangen/get-tarball": "^3.0.7",
-    "@storybook/codemod": "workspace:*",
     "@storybook/core": "workspace:*",
     "@types/semver": "^7.3.4",
-    "@yarnpkg/fslib": "2.10.3",
-    "@yarnpkg/libzip": "2.3.0",
     "chalk": "^4.1.0",
     "commander": "^6.2.1",
-    "cross-spawn": "^7.0.3",
-    "detect-indent": "^6.1.0",
-    "envinfo": "^7.7.3",
     "execa": "^5.0.0",
     "fd-package-json": "^1.2.0",
     "find-up": "^5.0.0",
     "fs-extra": "^11.1.0",
     "get-npm-tarball-url": "^2.0.3",
-    "giget": "^1.0.0",
-    "globby": "^14.0.1",
-    "jscodeshift": "^0.15.1",
-    "leven": "^3.1.0",
     "ora": "^5.4.1",
     "prettier": "^3.1.1",
     "prompts": "^2.4.0",
@@ -75,8 +63,6 @@
     "@types/prompts": "^2.0.9",
     "@types/util-deprecate": "^1.0.0",
     "boxen": "^7.1.1",
-    "slash": "^5.0.0",
-    "strip-ansi": "^7.1.0",
     "strip-json-comments": "^3.1.1",
     "typescript": "^5.3.2"
   },
