+++ code/lib/create-storybook/project.json
@@ -1,5 +1,5 @@
 {
-  "name": "toolbox",
+  "name": "create-storybook",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "targets": {
