+++ code/lib/create-storybook/src/NpmOptions.ts
@@ -1,0 +1,4 @@
+export type NpmOptions = {
+  skipInstall?: boolean;
+  installAsDevDependencies?: boolean;
+};
