+++ code/lib/create-storybook/src/detect.test.ts
