+++ code/lib/create-storybook/src/dirs.ts
