+++ code/lib/create-storybook/src/generators/NUXT/index.ts
