+++ code/lib/create-storybook/src/generators/SOLID/index.ts
