+++ code/lib/create-storybook/src/generators/SVELTE/index.ts
