+++ code/lib/create-storybook/src/generators/configure.test.ts
