+++ code/lib/create-storybook/src/scaffold-new-project.ts
