+++ code/lib/create-storybook/templates/angular/application/template-csf/.storybook/tsconfig.doc.json
