+++ code/lib/create-storybook/templates/angular/application/template-csf/.storybook/typings.d.ts
