+++ code/lib/create-storybook/templates/angular/library/template-csf/.storybook/tsconfig.json
