+++ code/lib/create-storybook/templates/angular/library/template-csf/.storybook/typings.d.ts
