+++ code/lib/create-storybook/templates/aurelia/template-csf/.storybook/tsconfig.json
