+++ code/lib/create-storybook/templates/aurelia/template-csf/.storybook/typings.d.ts
