+++ code/lib/toolbox/src/automigrate/fixes/angular-builders.ts
@@ -4,7 +4,7 @@ import chalk from 'chalk';
 import prompts from 'prompts';
 import type { Fix } from '../types';
 import { isNxProject } from '../../helpers';
-import { AngularJSON } from '../../generators/ANGULAR/helpers';
+import { AngularJSON } from '../../../../create-storybook/src/generators/ANGULAR/helpers';
 import type { JsPackageManager } from '@storybook/core/common';
 import { getFrameworkPackageName } from '../helpers/mainConfigFile';
